+++ src/createAnimatedComponent/index.tsx
