+++ src/reanimated2/component/FlatList.tsx
@@ -3,7 +3,7 @@ import React, { Component, forwardRef } from 'react';
 import type { FlatListProps, LayoutChangeEvent } from 'react-native';
 import { FlatList, StyleSheet } from 'react-native';
 import { AnimatedView } from './View';
-import createAnimatedComponent from '../../createAnimatedComponent/createAnimatedComponent';
+import createAnimatedComponent from '../../createAnimatedComponent';
 import type { ILayoutAnimationBuilder } from '../layoutReanimation/animationBuilder/commonTypes';
 import type { StyleProps } from '../commonTypes';
 import type { AnimateProps } from '../helperTypes';
