+++ src/reanimated2/component/Image.ts
@@ -1,6 +1,6 @@
 import type { ImageProps } from 'react-native';
 import { Image } from 'react-native';
-import createAnimatedComponent from '../../createAnimatedComponent/createAnimatedComponent';
+import createAnimatedComponent from '../../createAnimatedComponent';
 import { Component } from 'react';
 import type { AnimateProps } from '../helperTypes';
