+++ src/reanimated2/component/View.ts
@@ -1,6 +1,6 @@
 import type { ViewProps } from 'react-native';
 import { View } from 'react-native';
-import createAnimatedComponent from '../../createAnimatedComponent/createAnimatedComponent';
+import createAnimatedComponent from '../../createAnimatedComponent';
 import { Component } from 'react';
 import type { AnimateProps } from '../helperTypes';
